refac: Moved api calls to their own file
